Move api store folder
