+++ src/api/accounts.sagas.test.ts
@@ -1,6 +1,6 @@
 import { expectSaga } from 'redux-saga-test-plan';
 
-import { createJsonRpcRequest } from '@api/store/utils';
+import { createJsonRpcRequest } from '@api/utils';
 import { JsonRPCMethod } from '@config';
 import { fAccounts, fRequestOrigin } from '@fixtures';
